requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

PyYAML
pandas
pyarrow
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
timm
scikit-learn
Pillow
tqdm
